move motor-control out of pipettes
